backEnd/requirements.txt
PyMuPDF
google-genai
sentence_transformers
langchain
faiss
numpy